fix: #20 랜딩 페이지 분리 및 자동 로그인 로직 추가
